nginx::source now installs nginx ohai plugin
nginx[:binary] attribute will be set to nginx[:src_binary]
this attribute is important to notify the nginx ohai plugin where the
expected nginx binary is located. This binary is used to determine information
about the nginx installation.

additionally, there really is no reason for two possible binary paths. The
NGINX cookbook is a choose a type of installation - package or source

Signed-off-by: jtimberman <joshua@opscode.com>