Fix bug with nginx defaults not interpolating.
We need to use set rather than default since we are using interpolation of node
values. However we need to use set_unless so we don't clobber the node values.
This is less than ideal since set_unless means if values are changed here in
future, due to ohai_plugins storing these settings, nginx will not be
re-compiled.