Proper fix for COOK-1483
It turns out my original fix of changing set to "default" was correct. The real
reason it wasn't working sometimes is because the node['nginx']['dir'] variable
was being set in the attributes for the "default" recipe, not the source recipe.
Making sure the defaults are loaded before the source ones fixes all this
problems.