test-kitchen - append module names to attribute; configurations
The nginx::source recipe already does an include_recipe on the modules
in the attribute, so append the modules we want to include to that,
rather than include the recipe to prevent a compile flag race
condition (where the flags aren't added to the bash script).

The http_geoip_module isn't to be tested at this time, as the upstream
site is hostile to automated testing (403's multiple GET on the data
files). This recipe should probably be refactored into a separate
cookbook too, COOK-1952 is created to track this.

Test the ohai_plugin recipe first, since it is included by the others,
we'll want to know if a failure occurs early.