[COOK-2877] - depend on the runit cookbook
Runit is the default init style for debian/ubuntu systems, so we need
to specify a dependency on it, otherwise, e.g., chef-solo won't load
the resources/providers and the runit_service will fail in the source
recipe. While one could use the init init_style attribute value, this
doesn't "just work" for most users.