node attribute to disable commons conf
It would be great if we didnt have to config nginx, its config gets complex, and the templates dont allow all options out of the box. If you include a config over the top of the commons_config, it calls reload on every run. no one wants that. so a node attribute to disable config is the best way to get around this to enable the use of the cookbook for basic install and setup.