update ChefSpec syntax for SoloRunner vs Runner/ServerRunner
Prevents a lot of deprecation warnings, and we now depend on a ChefSpec
version higher than 4.1.x, which provides these methods.

Use of SoloRunner is much faster than ServerRunner, as we do not have to
wait for chef-zero to be populated each time. Functionally, these should
be the same from the nginx cookbook perspective.