Allow setting an empty string to prevent additional repos
Introduced in #119 (COOK-3030), the intent was to allow a user to
specify an attribute override to prevent the EPEL repo from being
added to an RHEL-based system.

However, using `nil` at any other level doesn't work as was expected,
due to opscode/chef#1507, which is the defaul behavior.

Instead of trying to add conditional logic here, we cast the attribute
to a string and evaluate if it is empty.

Adds a spec test for the branching logic.

fixes #243, closes #244, references #119