Use stable ChefDK to run tests
Instead of trying to maintain a dependency chain, use the current
released version of ChefDK to execute tests.

This is not a replacement for full integration testing.

Also remove the project's stale Gemfile, possibly to be replaced at a
later date with a subset of gems used for testing only.

Includes an update to rubocop's config and some style updates in tests.

Updates cookbook metadata in response to foodcritic warnings about
`suggests` and `*_url` fields.