Determine pidfile location correctly via a helper
Debian/Ubuntu changed their pidfile location, but the upstream packages haven’t changed. This leads to an evil if statement that doesn’t work in attribute form due to overrides. Move this logic to a helper and remove the attribute. This is probably the first of many things that will move to helpers instead of attributes.

Signed-off-by: Tim Smith <tsmith@chef.io>