Set LUAJIT_INC and LUAJIT_LIB environment variables when configuring nginx
These were previously exported after luajit was built,
but they were not available when the nginx configure
script is run. This might work, or it could allow nginx
to be built against Lua (not JIT) if it's installed.

This gives us a hook to build up the environment
necessary to build nginx (nginx_source_env), and
sets LUAJIT_INC and LUAJIT_LIB there.

Signed-off-by: Jim Ryan <jim@room118solutions.com>