Correct documentation for `rate_limiting_backoff` attribute
This attribute has been documented, according with its name, as the
rate-limiting backoff period for a zone. However, the `limit_req_zone`
directive it applies to has no such option[1] - the attribute is instead used to
supply the size of the rate-limit zone, the shared memory area that's used to
store the state used for rate monitoring.

Since we can't change the name of the attribute without breaking existing
client attribute customisations[2], at least document the attribute's effect and
note that the name is incorrect.

[1] http://nginx.org/en/docs/http/ngx_http_limit_req_module.html#limit_req_zone
[2] Or at least, I can't think of a sensible way to do so right now

Signed-off-by: Simon Coffey <simon.coffey@futurelearn.com>