Correctly compare nginx versions with multiple digits
The existing version comparison in this template used #to_f on the nginx
version string to test for versions >= 1.2. However, this fails in a number of
scenarios.

For example, the current default nginx version is '1.10.3', a later release
than 1.2. However, '1.10.3'.to_f => 1.1, for which the numerical test fails,
causing the real_ip_recursive directive to be incorrectly omitted.

This change introduces a version comparison helper, which is cribbed from
Gem::Version[1], simplified to remove the ability to compare version strings
with non-numeric parts. I chose to do this rather than use Gem::Version
itself, as it's not clear that nginx uses non-numeric version identifiers[2], so
I couldn't be confident that Gem::Version's interpretation of a version string
will match nginx's.

[1] https://github.com/rubygems/rubygems/blob/3f3337e6b789fc127f1aaa6e9895472446de0638/lib/rubygems/version.rb#L336-L361
[2] no non-numeric parts occur in the changelog's release history:
    http://nginx.org/en/CHANGES-1.12

Signed-off-by: Simon Coffey <simon.coffey@futurelearn.com>