Update the resource documentation to clarify extra variables properties
The extra variables properties don't do anything useful with the
default templates. @fernandoaleman explained it well.
"The initial thought was to provide a basic nginx.conf file that could be used out of the box with a few configuration options that could be modified. However, for those that needed a more customized config, you could create your own template in your wrapper cookbook that you could pass in via the conf_cookbook and conf_template properties and whatever conf_variables you provided would be available in your template, since Nginx has so many different config options available."

This change adds documentaion to the resource definitions to clarify the intent and use.